Moved CLI commands into cmd directory
