+++ cmd/cmd.go
@@ -1,0 +1,20 @@
+package cmd
+
+const (
+	// Headings
+	headingTarget = "# Targets"
+
+	// Subcommands
+	cmdTargets = "targets"
+	cmdLatest  = "latest"
+	cmdNext    = "next"
+	cmdTo      = "to"
+	cmdVersion = "version"
+
+	// Flag keys
+	flagDirectory   = "dir"
+	flagIgnoreEmpty = "ignore"
+	flagVersion     = "version"
+	flagForce       = "force"
+	flagType        = "type"
+)
