+++ codecov.yml
@@ -1,0 +1,2 @@
+ignore:
+  - cmd
